Add one to many relation, check userID in Token

+++ BacktestAPI/Controllers/BacktestController.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using BacktestAPI.Models;
-using BacktestAPI.Services.CalculationService;
+﻿using BacktestAPI.Models;
 using BacktestAPI.Services.TradeService;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -18,10 +14,12 @@ namespace BacktestAPI
     public class BacktestController : ControllerBase
     {
         private readonly ITradeService _tradeService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public BacktestController(ITradeService tradeService)
+        public BacktestController(ITradeService tradeService, IHttpContextAccessor httpContextAccessor)
         {
             _tradeService = tradeService;
+            _httpContextAccessor = httpContextAccessor;
         }
 
 
@@ -29,25 +27,22 @@ namespace BacktestAPI
         [HttpGet]
         public async Task<List<Trade>> GetAllTrades()
         {
+
             var allTrades = await _tradeService.GetAllTrades();
             return allTrades;
         }
 
-        // GET api/values/5
-        #nullable disable
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Trade>> GetSingleTrade(int id)
-        {
-            var result = await _tradeService.GetSingleTrade(id);
-            if (result is null)
-                return NotFound("Trade not found");
-            return Ok(result);
-        }
 
         // POST api/values
         [HttpPost]
         public async Task<ActionResult<List<Trade>>> AddTrade([FromBody] Trade trade)
         {
+            // get current userId from token
+            var userId = int.Parse(_httpContextAccessor.HttpContext.User.FindFirst("userId").Value);
+            if (userId != trade.UserId)
+            {
+                return BadRequest("You can add Trades only for your user!");
+            }
             var result = await _tradeService.AddTrade(trade);
             return Ok(result);
         }
@@ -56,6 +51,11 @@ namespace BacktestAPI
         [HttpPut("{id}")]
         public async Task<ActionResult<Trade>> UpdateTrade(int id, Trade request)
         {
+            var userId = int.Parse(_httpContextAccessor.HttpContext.User.FindFirst("userId").Value);
+            if (userId != request.UserId)
+            {
+                return BadRequest("You can update Trades only for your user!");
+            }
             var result = await _tradeService.UpdateTrade(id, request);
             if (result is null)
                 return NotFound("Trade not found!");
@@ -65,6 +65,11 @@ namespace BacktestAPI
         [HttpDelete("{id}")]
         public async Task<ActionResult<Trade>> DeleteTrade(int id, Trade request)
         {
+            var userId = int.Parse(_httpContextAccessor.HttpContext.User.FindFirst("userId").Value);
+            if (userId != request.UserId)
+            {
+                return BadRequest("You can delete Trades only for your user!");
+            }
             var result = await _tradeService.DeleteTrade(id,request);
             if (result is null)
                 return NotFound("Trade not found!");
